+++ index.html
@@ -227,6 +227,9 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
+  <script type="module" src="js/data.js"></script>
+  <script type="module" src="js/create-advert.js"></script>
+  <script type="module" src="js/util.js"></script>
 </body>
 </html>
