+++ js/data.js
@@ -1,0 +1,58 @@
+const TITLE_VALUES = [
+  'Каморка мечты',
+  'Белый дом',
+  'Уютное гнездо',
+  'Независимый подвал',
+  'Берлога туземца',
+];
+
+const TYPE_VALUES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const TIME_VALUES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES_VALUES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESC_VALUES = [
+  'Всё о чём можно мечтать',
+  'Подходит для последней ночи на планете',
+  'Вековая история за смешные деньги',
+  'Хороший номер, холодильник за доп плату',
+  'Результат вашего поиска',
+];
+
+const PHOTO_VALUES = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+const MAX_ROOMS = 3;
+const MAX_PRICE = 100000;
+const MAX_GUESTS = 5;
+
+export {
+  TITLE_VALUES, 
+  TYPE_VALUES, 
+  TIME_VALUES, 
+  FEATURES_VALUES, 
+  DESC_VALUES, 
+  PHOTO_VALUES, 
+  MAX_ROOMS, 
+  MAX_PRICE, 
+  MAX_GUESTS};
